Transverse Nucleon-to-Delta Analysis
The beam normal single spin asymmetry (Bn) is generated in the scattering of transversely polarized electrons from unpolarized nuclei. The asymmetry arises from the interference of the imaginary part of the two-photon exchange with the one-photon exchange amplitude. The Qweak experiment has made the first measurement of Bn in the production of the Δ(1232) resonance, using the Qweak apparatus in Hall-C at the Thomas Jefferson National Accelerator Facility. The final transverse asymmetry, corrected for backgrounds and beam polarization, is Bn = 43 ± 16 ppm at beam energy 1.16 GeV at an average scattering angle of about 8.3 degrees, and invariant mass of 1.2 GeV. The measured preliminary Bn agrees with a preliminary theoretical calculation. Bn for the Δ is the only known observable that is sensitive to the Δ elastic form-factors (γ*ΔΔ) in addition to the generally studied transition form-factors (γ*NΔ), but extracting this information will require